fix shout cmd
