+++ src/muscular-shout.js
@@ -1,15 +1,10 @@
 #!/usr/bin/env node
-const program   = require('commander');
-const getStdin  = require('get-stdin')
-const posejs    = require('./pose');
-const echoSdSlim    = require('./echo-sd-slim');
+const program  = require('commander');
+const getStdin = require('get-stdin')
+const posejs   = require('./pose');
+const shoutjs  = require('./shout');
 
-const _     = require('lodash')
-const fs    = require('fs');
-const tempy = require('tempy');
-const util  = require('util');
-const exec  = util.promisify(require('child_process').exec);
-
+const _ = require('lodash')
 const appRoot = `${__dirname}/..`
 
 let muscular_shout = () => {
@@ -30,29 +25,9 @@ program
     return
   }
 
-  let inputText = text ||
-    (await getStdin()) ||
-    muscular_shout();
+  let inputText = text || (await getStdin()) || muscular_shout();
 
-  let aa_text = posejs.action(options)
-  let aa_list = aa_text.split('\n')
-  let aa_height  = aa_list.length
-
-  let textCut = inputText.replace(new RegExp(`.{${(aa_height -1 ) - 2}}`,'g'),'$&\n')
-  let echo_sd_text = echoSdSlim.tatePrint(textCut)
-
-  //console.log({inputText,textCut,echo_sd_text})
-  let echo_sd_list = echo_sd_text.split('\n')
-
-  let echo_sd_list_centering = [
-    ..._.range(_.floor((aa_height - echo_sd_list.length) /2)).map(_=>''),
-    ...echo_sd_list
-  ]
-
-  let output = aa_list
-    .map((v,i)=>v + ' ' + (echo_sd_list_centering[i]||''))
-    .join('\n')
-
+  let output = shoutjs.action(inputText,options)
   console.log(output)
 })
 .parse(process.argv);
